Remove Frontend:: model namespace
The Frontend::* models existed only to layer presentation and query
methods on top of the base models, splitting every domain object across
two classes. Collapse that split by moving the logic into concerns and
the base models so controllers and views work with one class per concept.
