+++ app/controllers/frontend/recent_controller.rb
@@ -1,7 +1,7 @@
 module Frontend
   class RecentController < FrontendController
     def index
-      @events = Frontend::Event.recent(20).includes(:conference)
+      @events = Event.recent(20).includes(:conference)
 
       respond_to { |format| format.html }
     end
