+++ app/controllers/frontend/unpopular_controller.rb
@@ -5,14 +5,14 @@ module Frontend
     def index
       @year = params[:year].to_i
       @page = params[:page].to_i || 0
-      @firstyear = Frontend::Event.order('date ASC').limit(1).first.date.year
+      @firstyear = Event.order('date ASC').limit(1).first.date.year
       @events = if @year.zero?
-                  Frontend::Event.order('view_count ASC')
+                  Event.order('view_count ASC')
                                  .limit(PAGE_SIZE)
                                  .offset(@page * PAGE_SIZE)
                                  .includes(:conference)
                 else
-                  Frontend::Event.unpopular(@year)
+                  Event.unpopular(@year)
                                  .limit(PAGE_SIZE)
                                  .offset(@page * PAGE_SIZE)
                                  .includes(:conference)
