+++ app/graphql/resolvers/search_lectures.rb
@@ -5,7 +5,7 @@ class Resolvers::SearchLectures < GraphQL::Schema::Resolver
   argument :page, Integer, required: false
 
   def resolve(query:, page: 1)
-    results = Frontend::Event.query(query).page(page)
+    results = Event.query(query).page(page)
     @events = results.records.includes(recordings: :conference)
   end
 end
