--- app/models/frontend.rb
@@ -1,2 +1,0 @@
-module Frontend
-end
