+++ app/views/frontend/home/page_not_found.txt
@@ -1,0 +1,1 @@
+404
