+++ app/views/frontend/shared/_promoted.haml
@@ -1,7 +1,7 @@
 .promoted.dark.themed-banner{style: SiteSettings.current.promoted_banner_style}
   .titlebar
   .slider
-    - Frontend::Event.promoted(10).includes(:conference).each_with_index do |event, i|
+    - Event.promoted(10).includes(:conference).each_with_index do |event, i|
       .slide
         %a.item{href: event_path(slug: event.slug), 'class' => (i == 0 and 'active')}
           %img{src: h(event.thumb_url), alt: event.title, title: event.short_title}
